+++ rust/operator-binary/src/crd/mod.rs
@@ -9,7 +9,6 @@ use std::{collections::BTreeMap, str::FromStr};
 
 use affinity::get_affinity;
 use serde::{Deserialize, Serialize};
-use snafu::{OptionExt, Snafu};
 use stackable_operator::{
     commons::{
         affinity::StackableAffinity,
@@ -108,12 +107,6 @@ pub(crate) fn quantity_to_trino_bytes(
     Ok(format!("{bytes}B"))
 }
 
-#[derive(Snafu, Debug)]
-pub enum Error {
-    #[snafu(display("the role {role} is not defined"))]
-    CannotRetrieveTrinoRole { role: String },
-}
-
 #[versioned(
     version(name = "v1alpha1"),
     crates(
@@ -156,12 +149,10 @@ pub mod versioned {
         pub object_overrides: ObjectOverrides,
 
         // no doc - it's in the struct.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub coordinators: Option<super::TrinoCoordinatorRoleType>,
+        pub coordinators: super::TrinoCoordinatorRoleType,
 
         // no doc - it's in the struct.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub workers: Option<super::TrinoRoleType>,
+        pub workers: super::TrinoRoleType,
     }
 
     #[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
@@ -420,14 +411,22 @@ pub enum TrinoRole {
     Worker,
 }
 
+impl From<TrinoRole> for RoleName {
+    fn from(value: TrinoRole) -> Self {
+        RoleName::from_str(&value.to_string()).expect("a TrinoRole is a valid role name")
+    }
+}
+
+impl From<&TrinoRole> for RoleName {
+    fn from(value: &TrinoRole) -> Self {
+        RoleName::from_str(&value.to_string()).expect("a TrinoRole is a valid role name")
+    }
+}
+
 impl TrinoRole {
     pub fn listener_class_name(&self, trino: &v1alpha1::TrinoCluster) -> Option<ListenerClassName> {
         match self {
-            Self::Coordinator => trino
-                .spec
-                .coordinators
-                .to_owned()
-                .map(|coordinator| coordinator.role_config.listener_class),
+            Self::Coordinator => Some(trino.spec.coordinators.role_config.listener_class.clone()),
             Self::Worker => None,
         }
     }
@@ -505,29 +504,20 @@ impl v1alpha1::TrinoConfig {
 }
 
 impl v1alpha1::TrinoCluster {
-    /// Returns a reference to the role. Raises an error if the role is not defined.
-    pub fn role(&self, role_variant: &TrinoRole) -> Result<TrinoRoleType, Error> {
+    /// Returns the given role (both roles are required by the CRD).
+    pub fn role(&self, role_variant: &TrinoRole) -> TrinoRoleType {
         match role_variant {
-            TrinoRole::Coordinator => self
-                .spec
-                .coordinators
-                .to_owned()
-                .map(extract_role_from_coordinator_config),
+            TrinoRole::Coordinator => {
+                extract_role_from_coordinator_config(self.spec.coordinators.to_owned())
+            }
             TrinoRole::Worker => self.spec.workers.to_owned(),
         }
-        .with_context(|| CannotRetrieveTrinoRoleSnafu {
-            role: role_variant.to_string(),
-        })
     }
 
-    pub fn generic_role_config(&self, role: &TrinoRole) -> Option<&GenericRoleConfig> {
+    pub fn generic_role_config(&self, role: &TrinoRole) -> &GenericRoleConfig {
         match role {
-            TrinoRole::Coordinator => self
-                .spec
-                .coordinators
-                .as_ref()
-                .map(|c| &c.role_config.common),
-            TrinoRole::Worker => self.spec.workers.as_ref().map(|w| &w.role_config),
+            TrinoRole::Coordinator => &self.spec.coordinators.role_config.common,
+            TrinoRole::Worker => &self.spec.workers.role_config,
         }
     }
 
@@ -546,8 +536,8 @@ impl v1alpha1::TrinoCluster {
             .expect("the coordinator role name is a valid role name");
         self.spec
             .coordinators
+            .role_groups
             .iter()
-            .flat_map(|role| &role.role_groups)
             // Order rolegroups consistently, to avoid spurious downstream rewrites
             .collect::<BTreeMap<_, _>>()
             .into_iter()
@@ -642,6 +632,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
         "#;
@@ -661,6 +659,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
             tls:
@@ -682,6 +688,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
             tls:
@@ -701,6 +715,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
             tls:
@@ -725,6 +747,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
         "#;
@@ -744,6 +774,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
             tls:
@@ -765,6 +803,14 @@ mod tests {
         spec:
           image:
             productVersion: "481"
+          coordinators:
+            roleGroups:
+              default:
+                replicas: 1
+          workers:
+            roleGroups:
+              default:
+                replicas: 1
           clusterConfig:
             catalogLabelSelector: {}
             tls: